+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/MessageEntity/textsources/MentionTextSource.kt
@@ -18,7 +18,7 @@ private val String.withoutCommercialAt
  */
 data class MentionTextSource @RiskFeature(DirectInvocationOfTextSourceConstructor) constructor (
     override val source: String,
-    override val textSources: List<TextSource>
+    override val subsources: List<TextSource>
 ) : MultilevelTextSource {
     override val asMarkdownSource: String by lazy { source.mentionMarkdown() }
     override val asMarkdownV2Source: String by lazy { mentionMarkdownV2() }