+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/CommonSupergroupEventMessage.kt
@@ -4,16 +4,17 @@ import com.soywiz.klock.DateTime
 import dev.inmo.tgbotapi.types.MessageIdentifier
 import dev.inmo.tgbotapi.types.User
 import dev.inmo.tgbotapi.types.chat.abstracts.SupergroupChat
+import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.GroupEvent
 import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.SupergroupEvent
 import dev.inmo.tgbotapi.types.message.abstracts.SupergroupEventMessage
 
 @Deprecated("Renamed", ReplaceWith("CommonSupergroupEventMessage"))
-typealias SupergroupEventMessage = CommonSupergroupEventMessage
+typealias SupergroupEventMessage = CommonSupergroupEventMessage<*>
 
-data class CommonSupergroupEventMessage(
+data class CommonSupergroupEventMessage<T : SupergroupEvent>(
     override val messageId: MessageIdentifier,
     override val user: User,
     override val chat: SupergroupChat,
-    override val chatEvent: SupergroupEvent,
+    override val chatEvent: T,
     override val date: DateTime
-) : SupergroupEventMessage
+) : SupergroupEventMessage<T>