+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/abstracts/SupergroupEventMessage.kt
@@ -1,3 +1,5 @@
 package dev.inmo.tgbotapi.types.message.abstracts
 
-interface SupergroupEventMessage : GroupEventMessage
+import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.SupergroupEvent
+
+interface SupergroupEventMessage<T : SupergroupEvent> : GroupEventMessage<T>