+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/TextContent.kt
@@ -1,7 +1,6 @@
 package dev.inmo.tgbotapi.types.message.content
 
-import dev.inmo.tgbotapi.CommonAbstracts.TextSourcesList
-import dev.inmo.tgbotapi.CommonAbstracts.TextPart
+import dev.inmo.tgbotapi.CommonAbstracts.*
 import dev.inmo.tgbotapi.requests.abstracts.Request
 import dev.inmo.tgbotapi.requests.send.SendTextMessage
 import dev.inmo.tgbotapi.types.ChatIdentifier
@@ -15,13 +14,13 @@ import dev.inmo.tgbotapi.utils.internal.fullListOfSubSource
 import dev.inmo.tgbotapi.utils.internal.toMarkdownTexts
 
 data class TextContent(
-    val text: String,
-    /**
-     * Not full list of entities. This list WILL NOT contain [TextPart]s with [dev.inmo.tgbotapi.types.MessageEntity.textsources.RegularTextSource]
-     * @see [TextContent.fullEntitiesList]
-     */
-    val entities: List<TextPart> = emptyList()
-) : MessageContent {
+    override val text: String,
+    override val textEntities: List<TextPart> = emptyList()
+) : MessageContent, TextedInput {
+    @Deprecated("Has been renamed", ReplaceWith("textEntities"))
+    val entities: List<TextPart>
+        get() = textEntities
+
     override fun createResend(
         chatId: ChatIdentifier,
         disableNotification: Boolean,
@@ -83,4 +82,5 @@ data class TextContent(
  * Convert its [TextContent.entities] to list of [dev.inmo.tgbotapi.CommonAbstracts.TextSource]
  * with [dev.inmo.tgbotapi.types.MessageEntity.textsources.RegularTextSource]
  */
+@Deprecated("Useless due to the fact that currently every message contains full list of sources")
 fun TextContent.fullEntitiesList(): TextSourcesList = text.fullListOfSubSource(entities).map { it.source }