+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # TelegramBotAPI changelog
 
+## 0.30.4
+
+* `Common`:
+    * `Version`:
+        * `MicroUtils`: `0.3.1` -> `0.3.3`
+* `Core`:
+    * `MultilevelTextSource#textSources` has been safely renamed to `subsources`
+    * `TextContent#fullEntitiesList` has been deprecated
+    * Now `TextContent` implements `TextedInput`
+        * `TextContent#entities` has been deprecated
+    * `GroupEventMessage` now overrides `chatEvent` with type `GroupEvent`
+    * `SupergroupEventMessage` now overrides `chatEvent` with type `SupergroupEvent`
+    * Any `ChatEventMessage` now have generic type of its `chatEvent` (just like messages)
+* `Utils`:
+    * Old extensions related to chat events are deprecated:
+        * `Flow<ChatEventMessage<*>>#divideBySource`
+        * `Flow<ChatEventMessage<*>>#onlyChannelEvents`
+        * `Flow<ChatEventMessage<*>>#onlyGroupEvents`
+        * `Flow<ChatEventMessage<*>>#onlySupergroupEvents`
+    * A lot of extensions for `Flow<ChatEventMessage>` has been added:
+        * `FlowsUpdatesFilter#events`
+        * `FlowsUpdatesFilter#channelEvents`
+        * `FlowsUpdatesFilter#groupEvents`
+        * `FlowsUpdatesFilter#supergroupEvents`
+        * And a lot of other filters with specific types
+
 ## 0.30.3
 
 * `Common`: